Made following changes to make it work on ppc64le systems: 1. Created ppc64le specific Dockerfiles for pfe and performance components 2. Changed build.sh to disable codewind-appsody-extension for ppc64le

+++ script/build.sh
@@ -59,10 +59,14 @@ cp -r $DIR/docs ${SRC_DIR}/pfe/portal/
 # Copy the appsody and odo extension into portal. The zip file must have a version number e.g. codewind-appsody-extension-0.3.0.zip
 # in order for it to be accepted as a valid extension
 echo -e "\n+++   DOWNLOADING EXTENSIONS   +++\n";
-mkdir -p ${SRC_DIR}/pfe/extensions
-rm -f ${SRC_DIR}/pfe/extensions/codewind-appsody-extension-*.zip
+if [ "$ARCH" == "ppc64le" ]; then
+   echo "+++++   DISABLING APPSODY EXTENSION FOR PPC64LE"
+   mkdir -p ${SRC_DIR}/pfe/extensions
+   rm -f ${SRC_DIR}/pfe/extensions/codewind-appsody-extension-*.zip
+   curl -Lo ${SRC_DIR}/pfe/extensions/codewind-appsody-extension-9.9.9999.zip http://download.eclipse.org/codewind/codewind-appsody-extension/master/latest/codewind-appsody-extension-9.9.9999.zip
+fi
+
 rm -f ${SRC_DIR}/pfe/extensions/codewind-odo-extension-*.zip
-curl -Lo ${SRC_DIR}/pfe/extensions/codewind-appsody-extension-9.9.9999.zip http://download.eclipse.org/codewind/codewind-appsody-extension/master/latest/codewind-appsody-extension-9.9.9999.zip
 curl -Lo ${SRC_DIR}/pfe/extensions/codewind-odo-extension-9.9.9999.zip http://download.eclipse.org/codewind/codewind-odo-extension/master/latest/codewind-odo-extension-9.9.9999.zip
 
 # BUILD IMAGES
